LayoutTests/platform/glib/imported/w3c/web-platform-tests/editing/other/copy-paste-before-non-editable-expected.txt

Select and paste Selectsomething (like plain text) into the middle of this text node. 


FAIL Copy pasting in editable area should not lose data in non-editable area assert_equals: expected 3 but got 1
